scripts/.config/scripts/pyproject.toml
[project]
name = "dotfiles"
version = "0.1.0"
description = "My dotfiles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "python-decouple>=3.8",
    "shpyx>=0.0.32",
    "xlwings>=0.33.5",
]
